pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mtgv2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.9.0",
    "dagster==1.11.10",
    "dagster-dbt>=0.27.10",
    "dagster-dg-cli>=1.11.10",
    "dagster-webserver==1.11.10",
    "duckdb>=1.4.0",
    "pandas>=2.3.2",
    "psycopg2>=2.9.10",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "ruff>=0.13.2",
    "sqlalchemy>=2.0.43",
]

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "mtgv2"

[tool.pytest.ini_options]
pythonpath = ["src"]

[project.optional-dependencies]
postgres = ["psycopg2-binary"]
test = ["pytest", "duckdb"]
dev = ["psycopg2-binary", "pytest", "duckdb"]
